src/gateway-service/requirements.txt
Flask==2.2.2
psycopg2_binary==2.9.4
requests==2.28.1
python-dotenv>=0.19.2
authlib>=1.0
auth0-python==3.24.0
jwt==1.3.1